Refactor: Update dependencies and improve UI for document format selection
This commit updates several dependencies to their latest versions, including Gradle, Kotlin, and various AndroidX and Firebase libraries.

It also refines the UI for selecting document formats in the sales flow:
- `NewSalesTopBar.kt`: The color of the "Letter" and "Ticket" format options in the dropdown now visually indicates the currently selected format.
- `OpenFacturaRepository.kt`: Ensures `response.number` (Long) is correctly used.
- `Format50mmView.kt`: Minor vertical alignment adjustment for the timbre image.
- `NewSaleScreen.kt` and `SaleScreen.kt`: Pass the `isLetter` state to the respective top bars and view model methods.
- `SaleViewModel.kt`: Logs the `isLetter` state during sale creation and when the letter format is changed.

+++ app/src/main/java/com/blipblipcode/distribuidoraayl/data/repositiry/of/OpenFacturaRepository.kt
@@ -71,7 +71,7 @@ class OpenFacturaRepository @Inject constructor(
             }
             DocumentElectronic(
                 uri = uri,
-                number = response.number.toLong(),
+                number = response.number,
                 docType = DteType.INVOICE,
                 format = if(isLetter) DocFormat.LETTER else DocFormat.F80MM,
                 token = response.token,
